obj, texture 이름 받을수 있게 수정 디스크립터레이아웃, 디스크립터셋, 디스크립터 풀 vulkanModel로 이동,  VulkanRenderer  에서 삭제 uniform 버퍼 부분도 수정 uniform 버퍼 업데이트도 모델을 통하여 업데이트
VulkanRenderer 및 Scop 관련 함수 및 구조체 수정

- `BasicPSO.cpp`: `RState.cull_mode`와 `VIState.attributes`의 주석 제거
- `BmpLoader.cpp`: 디버그 출력 주석 처리
- `ObjMeshLoader.cpp`: `vertexCount` 타입을 `int`에서 `size_t`로 변경
- `Scop.cpp`, `Scop.h`: `update()` 호출 주석 처리, `prepare` 및 `createScopObject` 함수 인자 변경
- `Scop.vcxproj`, `Scop.vcxproj.filters`: `Types.h` 추가
- `ScopObject.cpp`, `ScopObject.h`: `CreateBoxMeshObject`, `CreateObjMeshObject`, `Initialize` 함수 인자 변경
- `VulkanModel.cpp`, `VulkanModel.h`: 소멸자, `Initialize`, `bind`, `EndCreateMesh`, `createDescriptorSetLayout`, `createDescriptorPool`, `createDescriptorSets`, `createPipelineLayout`, `updateUniformBuffer` 함수 추가 및 수정
- `VulkanPipeline.cpp`: 불필요한 주석 제거
- `VulkanRenderer`: 소멸자 및 여러 함수 수정 및 제거, `CreateTexture` 함수 추가
- `main.cpp`: `main` 함수 시그니처 변경, 명령줄 인수 처리 코드 추가, `scop->prepare` 호출 수정
- `Types.h`: `#pragma once` 추가, `Matrix.h` 포함, `ShaderData`, `UniformBuffer` 구조체 추가

+++ Scop/ObjMeshLoader.cpp
@@ -75,7 +75,7 @@ namespace scop {
 
 		// 중심축 이동 
 		// save triangle id
-		int vertexCount = vertices.size();
+		size_t vertexCount = vertices.size();
 		mymath::Vec3 center = sum / (float)vertexCount;
 
 		uint32_t triangleID = 0;
